+++ cmd/gen-commands.ts
@@ -195,6 +195,7 @@ export const commands = [
     "CmdDebugTogglePredictiveRender", "Debug: Toggle Predictive Rendering",
     "CmdDebugToggleRtl", "Debug: Toggle Rtl",
     "CmdToggleAntiAlias", "Toggle Anti-Alias Rendering",
+    "CmdListPrinters", "List Printers",
     "CmdToggleWindowsPreviewer", "Toggle Windows Previewer",
     "CmdToggleWindowsSearchFilter", "Toggle Windows Search Filter",
     "CmdNone", "Do nothing",