+++ src/Commands.h
@@ -73,7 +73,8 @@ enum {
     CmdDebugCorruptMemory = 384, CmdDebugCrashMe = 385, CmdDebugDownloadSymbols = 386,
     CmdDebugTestApp = 387, CmdDebugShowNotif = 388, CmdDebugStartStressTest = 389,
     CmdDebugTogglePredictiveRender = 390, CmdDebugToggleRtl = 391, CmdToggleAntiAlias = 392,
-    CmdToggleWindowsPreviewer = 393, CmdToggleWindowsSearchFilter = 394, CmdNone = 395,
+    CmdListPrinters = 393, CmdToggleWindowsPreviewer = 394, CmdToggleWindowsSearchFilter = 395,
+    CmdNone = 396,
 
     /* range for file history */
     CmdFileHistoryFirst,